[slapd] Allow use as Samba user database
In order for Samba to use openLDAP as its user database the Samba
schema must be included. Debian doesn't have a samba-ldap package, so
to avoid having to install samba server pckages on slapd hosts it is
necessary to include the schema directly in DebOps.

Samba does not use the standard LDAP password attribute, instead relying
on password hashes in the sambaLMPassword and sambaNTPassword
attributes. Samba service users must have read access to these
attributes, so a new Samba Server Account role has been included.

To ensure that the NT and LM password hashes are kept in sync with the
user's LDAP password, the slapd-contrib package needs to be installed
and the smbk5pwd overlay must be enabled. As this overlay is configured
to only apply to records with the sambaSamAccount objectClass, it is
safe to enable regardless of whether Samba LDAP integration is needed.